run/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gapp-run"
version = "0.1.0"
description = "gapp runtime — ASGI wrapper for credential mediation on Cloud Run"
requires-python = ">=3.10"
dependencies = [
    "starlette>=0.27",
    "google-auth>=2.0",
    "PyJWT>=2.0",
]

[tool.setuptools.packages.find]
include = ["gapp_run*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
